Add leave team button
