+++ src/group/styles.js
@@ -1,13 +1,5 @@
 import { defaultColor } from "~/shared/styles/colors";
-
-const button = {
-    fontSize: "28px",
-    width: "300px",
-    color: "white",
-    border: "none",
-    marginBottom: "20px",
-    boxShadow: "1.5px, 3px, 8px, -2px, #983745",
-}
+import buttons from "~/shared/styles/buttons";
 
 export default {
     buttonContainer: {
@@ -19,18 +11,5 @@ export default {
       textAlign: "center",
       marginBottom: "35px",
     },
-    leaveButton: [button, {
-        backgroundColor: "#DD2818",
-        pointer: "cursor",
-        ":hover": {
-            backgroundColor: "#C71111",
-        },
-    }],
-    inviteButton: [button, {
-        backgroundColor: "#1BB72D",
-        pointer: "cursor",
-        ":hover": {
-            backgroundColor: "#009933",
-        },
-    }],
+    buttons,
 }
